backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.44
psycopg[binary]==3.2.13
pydantic-settings==2.8.1
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
